requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
blobfile==3.0.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.8
deprecation==2.1.0
distro==1.9.0
et_xmlfile==2.0.0
filelock==3.18.0
fire==0.7.0
Flask==3.1.0
flask-cors==6.0.1
fsspec==2025.3.2
gotrue==2.12.3
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
llama_stack==0.2.2
llama_stack_client==0.2.2
lxml==5.3.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.1.3
openai==1.75.0
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
postgrest==1.1.1
prompt_toolkit==3.0.51
pyaml==25.1.0
pycryptodomex==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
PyYAML==6.0.2
realtime==2.5.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
setuptools==79.0.0
six==1.16.0
sniffio==1.3.1
storage3==0.12.0
StrEnum==0.4.15
supabase==2.16.0
supafunc==0.10.1
termcolor==3.0.1
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.14.1
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
